feat:消息发送

+++ toucan-shopping-modules/message/toucan-shopping-message-entity-vo/src/main/java/com/toucan/shopping/modules/message/entity/MessageBody.java
@@ -40,7 +40,7 @@ public class MessageBody {
     /**
      * 消息类型ID(用于消息历史)
      */
-    private Integer messageTypeId;
+    private Long messageTypeId;
 
     /**
      * 消息类型编码(用于消息历史)
